Assignment-data scraping and saving to database/requirements.txt
pandas
requests
html5lib
bs4
beautifulsoup4
mysql-connector-python